Fix large amount of size file cannot read

+++ Arduino15/packages/SPRESENSE/hardware/spresense/1.0.0/libraries/SDHCI/SDHCI.cpp
@@ -407,7 +407,7 @@ void File::flush() {
     fflush(_fd);
 }
 
-int File::read(void *buf, uint16_t nbyte) {
+int File::read(void *buf, size_t nbyte) {
   int ret;
 
   if (_fd) {
